nspawn: give container access to its loop device nodes
When running from a disk image copy the appropriate device nodes into
the container. This allows the container to manipulate the partition
table and mark the currently used partitions as successful.

Later it would be useful to copy in all of the device's partition nodes,
giving the container access to any lower priority inactive partitions.

For now this only works with --register=no because there is a second
hard-coded copy of container accessible device nodes hidden away in
src/core/cgroup.c which is used to configure the devices cgroup
controller. Some new plumbing will be required.